Finish 2d implementation
Note: this is going to be reverted since the idea doesn't work well and
the code has become too complex but a lot of effort has been put into
these changes. I want them to be at least committed such that it can be
eventually helpful in the future when I plan to revisit these ideas.

Signed-off-by: Steven Lang <steven.lang.mz@gmail.com>